we will need execvp() more places, move to util
